updated functionality to comply with database

+++ todolist-WebApp/app.js
@@ -1,6 +1,7 @@
 import express from "express";
 import bodyParser from "body-parser";
 import mongoose from "mongoose";
+import _ from "lodash";
 
 const app = express();
 
@@ -35,8 +36,15 @@ const default3 = new Item ({
 
 const defaultItems = [default1, default2, default3];
 
-app.get("/", async (req, res) =>  {
+// new list schema
+const listsSchema = {
+  name: String,
+  items: [itemsSchema]
+};
 
+const List = mongoose.model("List", listsSchema);
+
+app.get("/", async (req, res) =>  {
   const currentDate = new Date();
   const options = {
     weekday: "long", 
@@ -71,7 +79,7 @@ app.get("/", async (req, res) =>  {
       });
       res.redirect("/");
     } else {
-      res.render("list", { date: formattedDate, newListItems: foundItems });
+      res.render("list", { date: formattedDate, newListItems: foundItems, listTitle: "Main" });
     }
   } catch (err) {
     console.error("Error fetching items:", err);
@@ -79,41 +87,89 @@ app.get("/", async (req, res) =>  {
   }
 });
 
-app.post("/", async (req, res) => {
+// using express route parameters to make multiple lists 
+app.get("/:listName", async (req, res) => {
+  const listName = _.lowerCase(req.params.listName);
 
+  try {
+    const foundList = await List.findOne({ name: listName });
+
+    if (!foundList) {
+      const list = new List({
+        name: listName,
+        items: defaultItems
+      });
+    
+      await list.save();
+      res.redirect("/" + listName);
+    } else {
+      res.render("list", { date: formattedDate, newListItems: foundList.items, listTitle: listName });
+    }
+    
+  } catch (err) {
+    console.error("Error while accessing database:", err);
+    res.status(500).send("Error while accessing database.");
+  }
+});
+
+app.post("/", async (req, res) => {
   const itemName = req.body.newItem;
+  const listName = req.body.list;
 
   const item = new Item({
     name: itemName
   });
 
-  // after we save our item, we redirect so that we can FIND all items and render them on screen
-  try {
-    await item.save();
-    res.redirect("/");
-  } catch (err) {
-    console.error("Error saving item:", err);
-    res.status(500).send("Error saving item.");
+  if (listName === "Main") {
+    // after we save our item, we redirect so that we can FIND all items and render them on screen
+    try {
+      await item.save();
+      res.redirect("/");
+    } catch (err) {
+      console.error("Error saving item:", err);
+      res.status(500).send("Error saving item.");
+    }
+  } else {
+    try {
+      const foundList = await List.findOne({ name: listName });
+      foundList.items.push(item);
+      await foundList.save();
+      res.redirect("/" + listName);
+    } catch (err) {
+      console.error("Error while accessing database:", err);
+      res.status(500).send("Error while accessing database.");
+    }
   }
 });
 
 app.post("/delete", async (req, res) => {
-
   const itemToDelete = req.body.deleteItem;
+  const listName = req.body.listName;
 
-  try {
-    await Item.findByIdAndRemove(itemToDelete);
-    res.redirect("/");
-  } catch (err) {
-    console.error("Error in deleting item:", err);
-    res.status(500).send("Error deleting item.");
+  if (listName === "Main") {
+    try {
+      await Item.findByIdAndRemove(itemToDelete);
+      res.redirect("/");
+    } catch (err) {
+      console.error("Error in deleting item:", err);
+      res.status(500).send("Error deleting item.");
+    }
+  } else {
+    // efficient way of removing an item from a Mongo Array
+    try {
+      const foundList = await List.findOneAndUpdate({ name: listName }, { $pull: { items: { _id: itemToDelete }}});
+      res.redirect("/" + listName);
+    } catch (err) {
+      console.error("Error in deleting item from list:", err);
+      res.status(500).send("Error deleting item from list.");
+    }
   }
 });
 
-app.get("/about", function(req, res){
+app.get("/about", (req, res) => {
   res.render("about");
 });
 
-app.listen(3000, function() {
+app.listen(3000, () => {
   console.log("Server started on port 3000");
 });
